Finished modularizing applicant and employer register
Created separate partial html files for each of the categories in the registration page and reused the components in both the employer and applicant register pages. The page will take a bit of a loading impact, but at least maintance and reusability will be a thing. We can also now animate the segments. :o